Refactoring getTokenInstance
